# api-spec.yaml
openapi: 3.0.1
info:
  title: CirendsAPI
  version: "1.0"
servers:
  - url: https://cirends.runasp.net/api
  - url: http://localhost:5000/api
...
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      description: "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
...
paths:
...
  /Activities:
    get:
      tags: [Activities]
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                - { id: 1, name: "Kelionė", description: "Aprašymas", startDate: "2025-01-10T00:00:00Z", endDate: "2025-01-12T00:00:00Z" }
        "401": { description: Unauthorized }
    post:
      tags: [Activities]
      requestBody:
        required: true
        content:
          application/json:
            example: { name: "Kelionė", description: "Aprašymas", startDate: "2025-01-10T00:00:00Z", endDate: "2025-01-12T00:00:00Z", location: "Druskininkai" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              example: { id: 1, name: "Kelionė", startDate: "2025-01-10T00:00:00Z" }
        "400": { description: Bad Request }
        "401": { description: Unauthorized }
  /Activities/{id}:
    get:
      tags: [Activities]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example: { id: 1, name: "Kelionė", participants: [], tasks: [] }
        "400": { description: Bad Request }
        "404": { description: Not Found }
        "401": { description: Unauthorized }
    put:
      tags: [Activities]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            example: { name: "Kelionė (atnaujinta)" }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "404": { description: Not Found }
        "401": { description: Unauthorized }
        "403": { description: Forbidden }
    delete:
      tags: [Activities]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "404": { description: Not Found }
        "401": { description: Unauthorized }
        "403": { description: Forbidden }
...
  /Activities/{id}/participants/{userId}:
    delete:
      tags: [Activities]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: userId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "403": { description: Forbidden }
        "404": { description: Not Found }
        "401": { description: Unauthorized }
...